+++ lib/config/logging/navigation_logger.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:todo_list/config/logging/logger.dart';
-import 'package:todo_list/features/todo/presentation/utility/todo_action.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/utility/todo_action.dart';
 
 class NavigationLogger extends NavigatorObserver {
   @override
