+++ lib/features/todo/domain/bloc/todo_list_event.dart
@@ -1,10 +1,6 @@
-import 'package:equatable/equatable.dart';
 import 'package:todo_list/features/todo/domain/entity/todo.dart';
 
-abstract class TodoEvent extends Equatable {
-  @override
-  List<Object?> get props => [];
-}
+abstract class TodoEvent {}
 
 class LoadTodos extends TodoEvent {}
 
@@ -12,27 +8,18 @@ class AddTodoEvent extends TodoEvent {
   final Todo todo;
 
   AddTodoEvent(this.todo);
-
-  @override
-  List<Object?> get props => [todo];
 }
 
 class UpdateTodoEvent extends TodoEvent {
   final Todo todo;
 
   UpdateTodoEvent(this.todo);
-
-  @override
-  List<Object?> get props => [todo];
 }
 
 class DeleteTodoEvent extends TodoEvent {
   final int id;
 
   DeleteTodoEvent(this.id);
-
-  @override
-  List<Object?> get props => [id];
 }
 
 class ToggleVisibilityMode extends TodoEvent {}
