+++ lib/features/todo/domain/bloc/todo_list_state.dart
@@ -1,10 +1,6 @@
-import 'package:equatable/equatable.dart';
 import 'package:todo_list/features/todo/domain/entity/todo.dart';
 
-sealed class TodoState extends Equatable {
-  @override
-  List<Object> get props => [];
-}
+sealed class TodoState {}
 
 class TodoInitial extends TodoState {}
 
@@ -14,16 +10,10 @@ class TodoLoaded extends TodoState {
   final List<Todo> todos;
 
   TodoLoaded(this.todos);
-
-  @override
-  List<Object> get props => [todos];
 }
 
 class TodoError extends TodoState {
   final String message;
 
   TodoError(this.message);
-
-  @override
-  List<Object> get props => [message];
 }
