+++ lib/features/todo/presentation/todo_all/widgets/done_todo_count_widget.dart
@@ -1,14 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:todo_list/features/todo/domain/bloc/todo_bloc.dart';
-import 'package:todo_list/features/todo/domain/bloc/todo_state.dart';
+import 'package:todo_list/features/todo/domain/bloc/todo_list_bloc.dart';
+import 'package:todo_list/features/todo/domain/bloc/todo_list_state.dart';
 
 class DoneTodoCountWidget extends StatelessWidget {
   const DoneTodoCountWidget({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return BlocBuilder<TodoBloc, TodoState>(
+    return BlocBuilder<TodoListBloc, TodoState>(
       builder: (context, state) {
         switch (state) {
           case TodoInitial _:
