+++ lib/features/todo/presentation/todo_single/bloc/single_todo_state.dart
@@ -1,0 +1,7 @@
+import 'package:todo_list/features/todo/domain/entity/todo.dart';
+
+class SingleTodoState {
+  final Todo todo;
+
+  const SingleTodoState({required this.todo});
+}
