+++ lib/features/todo/presentation/todo_single/screens/single_todo_screen.dart
@@ -1,0 +1,63 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:todo_list/core/ui/custom_icon_button.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/bloc/single_todo_cubit.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/utility/todo_action.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/widgets/todo_content_text_field.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/widgets/todo_deadline_selection_tile.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/widgets/todo_delete_button.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/widgets/todo_priority_selection_tile.dart';
+import 'package:todo_list/features/todo/presentation/todo_single/widgets/todo_save_button.dart';
+
+class TodoScreen extends StatelessWidget {
+  final TodoAction action;
+  const TodoScreen({super.key, required this.action});
+
+  @override
+  Widget build(BuildContext context) {
+    return BlocProvider(
+      create: (context) => SingleTodoCubit(action: action),
+      lazy: false,
+      child: Scaffold(
+        body: CustomScrollView(
+          slivers: [
+            SliverAppBar(
+              elevation: 0,
+              scrolledUnderElevation: 5,
+              leading: CustomIconButton(
+                icon: Icons.close,
+                onPressed: () => Navigator.of(context).pop(),
+                color: Theme.of(context).colorScheme.onSurface,
+              ),
+              actions: const [TodoSaveButton()],
+              pinned: true,
+            ),
+            const SliverPadding(
+              padding: EdgeInsets.only(top: 8, right: 16, left: 16),
+              sliver: SliverToBoxAdapter(
+                child: Column(
+                  children: [
+                    TodoContentTextField(),
+                    TodoPrioritySelectionTile(),
+                    Divider(height: 0),
+                    TodoDeadlineSelectionTile(),
+                  ],
+                ),
+              ),
+            ),
+            const SliverToBoxAdapter(child: Divider()),
+            const SliverPadding(
+              padding: EdgeInsets.only(bottom: 8, right: 16, left: 16),
+              sliver: SliverToBoxAdapter(
+                child: Align(
+                  alignment: Alignment.centerLeft,
+                  child: TodoDeleteButton(),
+                ),
+              ),
+            ),
+          ],
+        ),
+      ),
+    );
+  }
+}
