--- lib/features/todo/presentation/utility/todo_result.dart
@@ -1,17 +1,0 @@
-import 'package:todo_list/features/todo/domain/entity/todo.dart';
-
-sealed class TodoResult {}
-
-final class CreatedTodo extends TodoResult {
-  final Todo todo;
-
-  CreatedTodo({required this.todo});
-}
-
-final class EditedTodo extends TodoResult {
-  final Todo? todo;
-
-  EditedTodo({required this.todo});
-}
-
-final class DeletedTodo extends TodoResult {}
