+++ pubspec.yaml
@@ -9,7 +9,6 @@ environment:
 dependencies:
   bloc: ^8.1.4
   dio: ^5.4.3+1
-  equatable: ^2.0.5
   flutter:
     sdk: flutter
   flutter_bloc: ^8.1.6
